\docType{data}
\name{hg19pvalues}
\alias{hg19pvalues}
\title{p-values from human RNA-Seq data with two conditions, four samples}
\format{a \code{matrix} with p-values from metaseqR2 supported tests.}
\source{
    Giakountis et al. (https://doi.org/10.1016/j.celrep.2016.05.038)
}
\description{
    This data set contains p-values calculated with each
    of the supported statistical testing algorithms in
    metaseqR2 for 1000 genes. The purpose of this matrix is
    to demonstrate the p-value combination methods as well as
    be used for a playground for other such methods and with
    other metaseqR2 facilities.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
