\docType{data}
\name{libsizeListMm9}
\alias{libsizeListMm9}
\title{Mouse RNA-Seq data with two conditions, four samples}
\format{a named \code{list} with library sizes.}
\source{
    ENCODE (http://genome.ucsc.edu/encode/)
}
\description{
    The library size list for \code{mm9GeneCounts}. See the
    data set description.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
